% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{areaAttributes}
\alias{areaAttributes}
\title{helper function to provide area properties}
\usage{
areaAttributes(
  size = 0.5,
  linetype = "solid",
  color = NA,
  fillColor = "red",
  alpha = 0.25
)
}
\arguments{
\item{size}{width of the line around the area}

\item{linetype}{linetype (solid, dotted, etc) of the line around the area}

\item{color}{color of the line around the area}

\item{fillColor}{color of the area}

\item{alpha}{alpha ('see through' factor) of the area}
}
\value{
a list object of parameters
}
\description{
helper function to provide area properties
}
