% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{areaDefinition}
\alias{areaDefinition}
\title{helper function to generate an area definition}
\usage{
areaDefinition(x = NA, y = NA)
}
\arguments{
\item{x}{x-coordinates of the 'corners' of the area}

\item{y}{y-coordinates of the 'corners' of the area

returns a data.frame with columns x & y}
}
\description{
helper function to generate an area definition
}
\note{
areaBlockExample <- areaDefinition(x = c(5,10,10,5),
                                         y = c(10,10,20,20))
}
