% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{areaMarks}
\alias{areaMarks}
\title{to add closed (eg colored) areas to a list of ggplot objects}
\usage{
areaMarks(graphs, areas = areaDefinition(), areaAttributes = areaAttributes())
}
\arguments{
\item{graphs}{list of ggplot-objects to which the areas have to be added
Note: MUST be a list}

\item{areas}{list (!) of data.frames of x and y columns defining the area
to be drawn in the ggplot objects. Example: 
areaBlockExample <- data.frame(x = c(5,10,10,5), y = c(10,10,20,20))}

\item{areaAttributes}{list which defines the type of area to be used for
the areas}
}
\value{
a list of ggplot objects
}
\description{
to add closed (eg colored) areas to a list of ggplot objects
}
\note{
due to the graphs parameter you can use:
 list(ggplot objects) %>% graphAdjust() %>% areaMarks()
}
