% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{blobToVector}
\alias{blobToVector}
\title{does the reverse of the vectorToBlob function: splits a character/raw vector
 into seperate elements (after decompressing)}
\usage{
blobToVector(
  blobData = NA,
  collapseChar = ";",
  vectorClass = "integer",
  type = "gzip"
)
}
\arguments{
\item{blobData}{a character/raw vector to be splitted into seperate elements}

\item{collapseChar}{seperator to split dbData into seperate elements}

\item{vectorClass}{class of the seperate elements, default = "integer".
Other tested options are "numeric" and "character".}

\item{type}{character string, the type of compression. See ?memCompress for
details}
}
\value{
a vector of type vectorClass
}
\description{
does the reverse of the vectorToBlob function: splits a character/raw vector
 into seperate elements (after decompressing)
}
