% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{booleanToCharacter}
\alias{booleanToCharacter}
\title{Translates a logical vector into a character vector compromised of "F" and
 "T". All TRUE's become "T"'s, all FALSE's become "F"'s}
\usage{
booleanToCharacter(booleans)
}
\arguments{
\item{booleans}{logical vector to be translated}
}
\value{
character vector
}
\description{
Translates a logical vector into a character vector compromised of "F" and
 "T". All TRUE's become "T"'s, all FALSE's become "F"'s
}
\note{
NA and NAN give unpredictable results
}
