% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{characterToBoolean}
\alias{characterToBoolean}
\title{Translates a character vector compromised "F" and "T" into a logical vector
 of FALSE and TRUE. All "F"'s become FALSE's, all "T"'s become TRUE'S}
\usage{
characterToBoolean(characters)
}
\arguments{
\item{characters}{character vector to be translated}
}
\value{
logical vector
}
\description{
Translates a character vector compromised "F" and "T" into a logical vector
 of FALSE and TRUE. All "F"'s become FALSE's, all "T"'s become TRUE'S
}
\note{
NA and NAN give unpredictable results
}
