% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{clearPlot}
\alias{clearPlot}
\title{helper function to clear a plot area or to generate an empty ggplot object}
\usage{
clearPlot()
}
\value{
empty ggplot object
}
\description{
helper function to clear a plot area or to generate an empty ggplot object
}
