% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{combineDFSingleColumn}
\alias{combineDFSingleColumn}
\title{Takes a list of data.frame's and melts together the specified column from
 each data.frame. The resulting data.frame can be used for ...multi
 functions, eg statHistMultiple. Essentially this is a 'melt' function to
 join together data from different data.frame's}
\usage{
combineDFSingleColumn(
  data,
  column = 1,
  variableName,
  useWidth = 1,
  padChar = "0",
  variableClass = "character",
  variableColumn = "variable",
  valueColumn = "value"
)
}
\arguments{
\item{data}{the data to be used, should be a list of data.frames}

\item{column}{defines which column (must be a single one) from each of the
data.frame's is to be used to melt together.}

\item{variableName}{character vector, name(s)/identifier(s) of the different
data.frame's. If length of this argument is 1, then it is used together with
useWidth and padChar arguments. The full names of the elements of the list
will then be eg norm001, norm002, etc. If length is same as the length of
the list (of data.frame's), then the character vectors will be used as names
without adding numbers}

\item{useWidth}{'width' of the number to use together with the variableName
argument, see also the function stringr::str_pad which is used in this
function}

\item{padChar}{padding character to use together with useWidth and padChar
when using numbered identifiers, see also variableName and useWidth
arguments}

\item{variableClass}{character vector: defines the class for the variable
column of the resulting data.frame. Options are "character" (default), "integer"
and "numeric"}

\item{variableColumn}{character vector: defines the name of the variable
column of the resulting data.frame (identifier column for the list elements
from which the data was taken to melt together)}

\item{valueColumn}{character vector: defines the name of the value column of
the resulting data.frame}
}
\value{
a data.frame
}
\description{
Takes a list of data.frame's and melts together the specified column from
 each data.frame. The resulting data.frame can be used for ...multi
 functions, eg statHistMultiple. Essentially this is a 'melt' function to
 join together data from different data.frame's
}
