% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools.R
\name{confidenceIntervals}
\alias{confidenceIntervals}
\title{returns a confidence interval table of the specified data (mean, standard 
 deviation and associated statistical measures)}
\usage{
confidenceIntervals(data, column, removeNA = TRUE, alpha = 0.05)
}
\arguments{
\item{data}{the data to be used, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which column is to be used for calculating the
quantiles. Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to calculating the quantiles. @note this will remove warning messages and
errors}

\item{alpha}{(1-confidence level) for the upper and lower levels of the mean
and the standard deviation}
}
\value{
a data.frame with two columns: statistic and the values calculated
 from the data
}
\description{
returns a confidence interval table of the specified data (mean, standard 
 deviation and associated statistical measures)
}
