% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{convertDFToDB}
\alias{convertDFToDB}
\title{Converts a data.frames to a single data.frame where each data.frame from the
 list is a row of blobs or character separated strings}
\usage{
convertDFToDB(
  dfData,
  columnNames = colnames(dfData[[1]]),
  collapseChar = ";",
  formatNumbers = FALSE,
  toBlob = TRUE,
  type = "gzip",
  ...
)
}
\arguments{
\item{dfData}{the data to be converted: a list of data.frames; they can be
different number of rows, but columns specified in columnNames must be
present}

\item{columnNames}{character vector specifying which columns from the
data.frame(s) to convert}

\item{collapseChar}{the character to be used as seperator in the end result.
Note: this seperator should not be present in the dataVector (especially
after transformation) as it will give problems when using the DBToVector
function.}

\item{formatNumbers}{boolean specifying whether the dataVector should be
formatted using the base::format function}

\item{toBlob}{should the data be converted to a raw vector or not}

\item{type}{character string, the type of compression. See ?memCompress for
details}

\item{...}{further arguments to be passed onto to base::format, see
?base::format for more information. Ignored when formatNumbers == FALSE}
}
\value{
a raw vector
}
\description{
Converts a data.frames to a single data.frame where each data.frame from the
 list is a row of blobs or character separated strings
}
