% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{copyDBtoMemory}
\alias{copyDBtoMemory}
\title{copies contents of a sqlite database to another one (default is into
 :memory:)}
\usage{
copyDBtoMemory(
  dbFrom,
  dbTo = ":memory:",
  noWarnings = TRUE,
  copyTables = NA,
  doNotCopyTables = "sqlite_sequence"
)
}
\arguments{
\item{dbFrom}{database from which tables are to be copied, must be open when
calling this function}

\item{dbTo}{database to which tables are to be copied}

\item{noWarnings}{suppress warnings while copying data, default = TRUE}

\item{copyTables}{character vector specifying which tables to copy (default
NA, copies all tables)}

\item{doNotCopyTables}{character vector that specifies which tables are not
to be copied (default = 'sqlite_sequence'). Overrules the copyTables
argument. If NA then this parameter is ignored}
}
\value{
an in memory database database, note that this database is returned
 in an 'open state'
}
\description{
copies contents of a sqlite database to another one (default is into
 :memory:)
}
