% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{copyTable}
\alias{copyTable}
\title{Copies a single table from one database to another (only sqlite type
 databases have been tested). Internal function}
\usage{
copyTable(dbFrom, dbTo, tableName, noWarnings = TRUE)
}
\arguments{
\item{dbFrom}{database from which tables are to be copied, must be open when
calling this function}

\item{dbTo}{database to which tables are to be copied, must be open when
calling this function}

\item{tableName}{character vector specifying which table to copy}

\item{noWarnings}{suppress warnings while copying data, default = TRUE}
}
\value{
nothing
}
\description{
Copies a single table from one database to another (only sqlite type
 databases have been tested). Internal function
}
\note{
internal function
}
