% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dateToInteger}
\alias{dateToInteger}
\title{converts date format to integer numbers}
\usage{
dateToInteger(aDate, origin = as.Date("1970-01-01"))
}
\arguments{
\item{aDate}{Date formatted dates to be converted integer}

\item{origin}{date vector specifying the origin date}
}
\value{
converted dates in integer format
}
\description{
converts date format to integer numbers
}
