% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_close}
\alias{db_close}
\title{Wrapper around pool::pooClose():  closes an open database
(normally opened earlier via eg db_open())}
\usage{
db_close(db, suppressErrors = FALSE)
}
\arguments{
\item{db}{database access 'handle' to be closed}

\item{suppressErrors}{logical (default = FALSE). For the somewhat unique
situation that a database is possibly already closed. This happens eg when
working with 'in memory' databases as is possible with SQLite; the closing
of this type of database may already have happened when for instance a shiny
application closes/shuts down}
}
\description{
Wrapper around pool::pooClose():  closes an open database
(normally opened earlier via eg db_open())
}
