% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_columnInfo}
\alias{db_columnInfo}
\title{get the classes of the different columns in a database table}
\usage{
db_columnInfo(db, tableName)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}
}
\value{
a data.frame with columns name & type (= class)
}
\description{
get the classes of the different columns in a database table
}
