% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_createDataframe}
\alias{db_createDataframe}
\title{creates an empty data.frame base on what's in a database table}
\usage{
db_createDataframe(db, tableName, dbType = "SQLite")
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}

\item{dbType}{currently only SQLite is a valid value}
}
\value{
data.frame
}
\description{
creates an empty data.frame base on what's in a database table
}
\note{
dates come out of database (SQLite) as TEXT field, manual conversion
 to POSIXCT is needed afterwards
}
