% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_deleteRow}
\alias{db_deleteRow}
\title{removes rows from a table in the database specified by the 'where' argument}
\usage{
db_deleteRow(db, tableName, where = NA)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}

\item{where}{allows for conditional retrieving, should be in the form
of a character vector specifying the conditions and boolean operators, eg:
c("age = 45", " AND ", "gender = 'M'). if this parameter is NA, then nothing
happens}
}
\value{
nothing
}
\description{
removes rows from a table in the database specified by the 'where' argument
}
