% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_deleteTable}
\alias{db_deleteTable}
\title{delete a table in a database}
\usage{
db_deleteTable(db, tableName)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}
}
\value{
nothing
}
\description{
delete a table in a database
}
\note{
this action cannot be undone
}
