% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_getColumn}
\alias{db_getColumn}
\title{gets specified columns from a table in the database, allows for where
 statements}
\usage{
db_getColumn(db, tableName, column = NA, where = NA)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}

\item{column}{if NA then all columns are retrieved, otherwise should be
a character vector specifying the names of the columns to be retrieved}

\item{where}{allows for conditional retrieving, should be in the form
of a character vector specifying the conditions and boolean operators, eg:
c("age = 45", " AND ", "gender = 'M')}
}
\value{
data.frame
}
\description{
gets specified columns from a table in the database, allows for where
 statements
}
