% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_getLinkColumns}
\alias{db_getLinkColumns}
\title{retrieves the specified columns from the 'right' table when the IDs from
 the 'left' table are provided or vice versa. So gets the  columns from the
 other table that where the ID's are linked/matched}
\usage{
db_getLinkColumns(
  db,
  linkedTableName,
  hasPrimaryKey = FALSE,
  tableName,
  columns = NA,
  primaryKey = defaultPrimaryKeyName,
  left = TRUE,
  leftID = NA,
  rightID = NA,
  IDs = NA
)
}
\arguments{
\item{db}{database access 'handle'}

\item{linkedTableName}{name of the 'link' table to use}

\item{hasPrimaryKey}{specifies if the 'link' table has a primary key,
default is FALSE}

\item{tableName}{specfies the name of the table from where to get
the columns via the ID's specified by the 'link' table}

\item{columns}{specifies which columns to retrieve from the table
specified by tableName}

\item{primaryKey}{specifies the name of the primary key of the table
specified by tableName. These will be used to retrieve the right rows
from the table}

\item{left}{if TRUE then the provided ID's are from the 'left'
table and the corresponding ID's from the 'right' table are retrieved. If
FALSE then it's vice versa}

\item{leftID}{name of the ID column in the left table to be used,
if NA, the column 1 is used}

\item{rightID}{name of the ID column in the right table to be used,
if NA, the column 2 is used}

\item{IDs}{integer vector specifying the IDs that need to be
'matched'. If an ID is not present, then it will not generate anything}
}
\value{
integer vector
}
\description{
retrieves the specified columns from the 'right' table when the IDs from
 the 'left' table are provided or vice versa. So gets the  columns from the
 other table that where the ID's are linked/matched
}
\note{
if not specified, then leftID is always column 1, rightID is always
 column 2
}
