% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_getSQLQuery}
\alias{db_getSQLQuery}
\title{to perform general SQL select like statements}
\usage{
db_getSQLQuery(db, SQLString, doNotChangePool = FALSE)
}
\arguments{
\item{db}{database access 'handle'}

\item{SQLString}{the SQL query to be performed}

\item{doNotChangePool}{boolean that sets whether the pool has to be locked/
released inside this function. Usually this needs to be done, so default
is FALSE. Set to TRUE if this locking/release is done outside of the
function. Note that there is a sort of bug when working with the
pool::dbWriteTable : when a sql constraint like UNIQUE is defined for the
table used, this may result in a locked database upon error. It's advisable
to set doNotChangePool to TRUE to prevent this from happening}
}
\value{
result from the SQL query, usually a data.frame
}
\description{
to perform general SQL select like statements
}
