% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_getTable}
\alias{db_getTable}
\title{gets data from a table in the database}
\usage{
db_getTable(db, tableName = NA, SQLString = NA)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}

\item{SQLString}{allows for more specifics actions, if tableName == NA,
then SQLString is used as such, otherwise it will be used in the form of
SQLString + tableName}
}
\value{
data.frame
}
\description{
gets data from a table in the database
}
