% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_ncol}
\alias{db_ncol}
\title{find the number of columns in a table of an (open) database}
\usage{
db_ncol(db, tableName)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}
}
\value{
number of columns in the table
}
\description{
find the number of columns in a table of an (open) database
}
