% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_nrow}
\alias{db_nrow}
\title{find the number of rows in a table of an (open) database}
\usage{
db_nrow(db, tableName)
}
\arguments{
\item{db}{database access 'handle'}

\item{tableName}{name of the table}
}
\value{
number of rows in the table
}
\description{
find the number of rows in a table of an (open) database
}
