% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{db_tbl_def}
\alias{db_tbl_def}
\title{gathers info on all tables in the database}
\usage{
db_tbl_def(db)
}
\arguments{
\item{db}{database access 'handle'}
}
\value{
a data.frame with the following columns: name, nrows, info
every entry in the column info contains a data.frame with the names
of the columns in the table and their type (see also db_columnInfo)
}
\description{
gathers info on all tables in the database
}
