% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{DBtoVector}
\alias{DBtoVector}
\title{does the reverse of the vectorToDB function: splits a single element
 character vector into seperate elements}
\usage{
DBtoVector(dbData = NA, collapseChar = ";", vectorClass = "integer")
}
\arguments{
\item{dbData}{single element character vector to be splitted into seperate
elements}

\item{collapseChar}{seperator to split dbData into seperate elements}

\item{vectorClass}{class of the seperate elements, default = "integer".
Other tested options are "numeric" and "character".}
}
\value{
a vector of type vectorClass
}
\description{
does the reverse of the vectorToDB function: splits a single element
 character vector into seperate elements
}
