% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\docType{data}
\name{defaultPrimaryKeyName}
\alias{defaultPrimaryKeyName}
\title{'global' variable: default primary key name}
\format{
An object of class \code{character} of length 1.
}
\usage{
defaultPrimaryKeyName
}
\description{
'global' variable: default primary key name
}
\note{
in future: remove from export
}
\keyword{datasets}
