% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{fitDF}
\alias{fitDF}
\title{creates a fit table data.frame with the columns predictorColumn,
 ResponseColumn, Calculated Predictor, Recovery of predictor}
\usage{
fitDF(
  fitTable,
  predictorColumn = 1,
  responseColumn = 2,
  weights = NULL,
  fit = fitLM(fitTable, predictorColumn = predictorColumn, responseColumn =
    responseColumn, weights = weights),
  removeInf = TRUE
)
}
\arguments{
\item{fitTable}{data.frame containing the data to be fitted}

\item{predictorColumn}{can be integer (column number) or character
(column name)}

\item{responseColumn}{can be integer (column number) or character
(column name)}

\item{weights}{default = NULL, otherwise must be integer to give weights to
the predictor values (see also function weightsVector) or a numeric
vector with length = number of rows in calibrationTable}

\item{fit}{model object (eg coming from lm() or fitLM())}

\item{removeInf}{Due to the nature of the calculations, the recovery
sometimes results in +Inf or -Inf (infinite values). If this option is TRUE
then these values are replaced by NA}
}
\value{
a data.frame witg columns: predictorColumn, ResponseColumn,
 Calculated Predictor, Recovery of predictor
}
\description{
creates a fit table data.frame with the columns predictorColumn,
 ResponseColumn, Calculated Predictor, Recovery of predictor
}
