% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{fitLM}
\alias{fitLM}
\title{wrapper around the lm() function to streamline it's use a bit}
\usage{
fitLM(fitTable, predictorColumn = 1, responseColumn = 2, weights = NULL)
}
\arguments{
\item{fitTable}{data.frame containing the data to be fitted}

\item{predictorColumn}{can be integer (column number) or character
(column name)}

\item{responseColumn}{can be integer (column number) or character
(column name)}

\item{weights}{default = NULL, otherwise must be integer to give weights to
the predictor values (see also function weightsVector) or a numeric
vector with length = number of rows in calibrationTable

@returns fit object of type "lm"
@export}
}
\description{
wrapper around the lm() function to streamline it's use a bit
}
