% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{formatDigitsWaiver}
\alias{formatDigitsWaiver}
\title{Function to convert numerical vectors into strings. This specific one can be
 used in stead of the function factories below if needed when no options are
 needed}
\usage{
formatDigitsWaiver(v)
}
\arguments{
\item{v}{numeric vector to be converted to a character vector}
}
\value{
a character vector
}
\description{
Function to convert numerical vectors into strings. This specific one can be
 used in stead of the function factories below if needed when no options are
 needed
}
