% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{formatMinimumDigits}
\alias{formatMinimumDigits}
\title{Function factory to be used to specify the minimum number of digits to be
 used in numbers. The function generates numbers as strings}
\usage{
formatMinimumDigits(digits)
}
\arguments{
\item{digits}{integer value that specifies the number of digits to be used
by the resulting function}
}
\value{
a function that will take a numeric vector as an argument and 
 returns a character vector of the numeric vector with the set minimum number
 of digits (see ?formatC for more info)
}
\description{
Function factory to be used to specify the minimum number of digits to be
 used in numbers. The function generates numbers as strings
}
