% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{graphCurves}
\alias{graphCurves}
\title{function to generate ggplot object (in form of a single item list) from a
 list of data.frames of data}
\usage{
graphCurves(tables, x, y, lineWidth = 0.5, xLimits = NA, yPercentage = FALSE)
}
\arguments{
\item{tables}{list of data.frames from which to generate the graphs
Note: MUST be a list}

\item{x}{character vector defining which column from the data.frames to use
for the x-axis}

\item{y}{character vector defining which column from the data.frames to use
for the y-axis}

\item{lineWidth}{defines the width of the lines in the graphs drawn from the
x,y coordinates}

\item{xLimits}{defines the range x coordinates c(minimum, maximum),
all coordinates outsude this range are not drawn}

\item{yPercentage}{defines if y-axis maximum should be set to 100 percent}
}
\value{
a (single element) list of ggplot objects
}
\description{
function to generate ggplot object (in form of a single item list) from a
 list of data.frames of data
}
\note{
due to the graphs parameter you can use:
 graphCurves(tables) %>% graphAdjust() %>% lineMarks() and similar pipes
}
