% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{graphCurvesCombine}
\alias{graphCurvesCombine}
\title{to generate a ggplot object (actually a list(ggplot object) ) which combines
the data of a number of tables}
\usage{
graphCurvesCombine(
  tables,
  x,
  y,
  lineWidth = 0.5,
  xLimits = NA,
  yPercentage = FALSE,
  combine = NA,
  colors = "black"
)
}
\arguments{
\item{tables}{list of data.frames from which to generate the graphs
Note: MUST be a list}

\item{x}{character vector defining which column from the data.frames to use
for the x-axis}

\item{y}{character vector defining which column from the data.frames to use
for the y-axis}

\item{lineWidth}{defines the width of the lines in the graphs drawn from the
x,y coordinates}

\item{xLimits}{defines the range x coordinates c(minimum, maximum),
all coordinates outsude this range are not drawn}

\item{yPercentage}{defines if y-axis maximum should be set to 100%}

\item{combine}{if NA then all tables are combined into one, otherwise it
has to be a integer vector defining which tables from the 'tables' list
to combine eg c(1,2,4)}

\item{colors}{character vector describing the colors to be used. If a single
color (default = "black") then all curves will be the same color, otherwise
must be character vector of same length as the combine vector}
}
\value{
a list of ggplot objects
}
\description{
to generate a ggplot object (actually a list(ggplot object) ) which combines
the data of a number of tables
}
