% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{histP}
\alias{histP}
\title{Plots a histogram with percentages in stead of frequencies on the y-axis}
\usage{
histP(data, breaks = "Sturges", ylab = "Percentage", ...)
}
\arguments{
\item{data}{data to be made into a histogram}

\item{breaks}{see ?graphics::hist}

\item{ylab}{label for y-axis, default is 'Percentage'}

\item{...}{further arguments to be passed onto the graphics::plot function}
}
\value{
a (base) plot
}
\description{
Plots a histogram with percentages in stead of frequencies on the y-axis
}
