% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{joinNamedDFList}
\alias{joinNamedDFList}
\title{joins together a named list of data.frame's with exact same columns. Adds
 a column to each data.frame in the list with the name of the list element}
\usage{
joinNamedDFList(listData)
}
\arguments{
\item{listData}{the named list of data.frame's}
}
\value{
a data.frame
}
\description{
joins together a named list of data.frame's with exact same columns. Adds
 a column to each data.frame in the list with the name of the list element
}
