% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{limits}
\alias{limits}
\title{helperfunction for limit line attributes}
\usage{
limits(
  yIntercept = c(0),
  color = c("blue"),
  linetype = c(1),
  size = c(1),
  alpha = c(0.5)
)
}
\arguments{
\item{yIntercept}{numeric vector where to place limit lines on the y-axis}

\item{color}{colors of the limit lines}

\item{linetype}{linetype (solid, dotted, etc) of the limit lines}

\item{size}{width of the limit lines}

\item{alpha}{alpha ('see through' factor) of the limit lines}
}
\value{
data.frame of the attributes
}
\description{
helperfunction for limit line attributes
}
\note{
Other parameters are either one per line or one for all

if no limits are wanted/needed, set at least one parameter to NA
}
