% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{lineAttributes}
\alias{lineAttributes}
\title{helper function to provide line properties}
\usage{
lineAttributes(size = 1, linetype = "solid", color = "black", alpha = 1)
}
\arguments{
\item{size}{width of the line}

\item{linetype}{linetype (solid, dotted, etc) of the line}

\item{color}{color of the line}

\item{alpha}{alpha ('see through' factor) of the line}
}
\value{
a list object of parameters
}
\description{
helper function to provide line properties
}
