% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{lineMarks}
\alias{lineMarks}
\title{to add horizontal or vertical lines to a list of ggplot objects}
\usage{
lineMarks(
  graphs,
  vlines = NA,
  hlines = NA,
  vlinesAttributes = linesMarkDefaults(),
  hlinesAttributes = linesMarkDefaults()
)
}
\arguments{
\item{graphs}{a ggplot object or a list of ggplot-objects to which the lines
have to be added}

\item{vLines}{a numeric vector defining where the vertical lines are to be
placed}

\item{hLines}{a numeric vector defining where the vertical lines are to be
placed}

\item{vLinesAttributes}{list which defines the type of line to be used for
vlines parameter}

\item{hLinesAttributes}{list which defines the type of line to be used for
hlines parameter}
}
\value{
a list of ggplot objects
}
\description{
to add horizontal or vertical lines to a list of ggplot objects
}
\note{
due to the graphs parameter you can use:
 list(ggplot objects) %>% graphAdjust() %>% lineMarks()
}
