% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{linesMarkDefaults}
\alias{linesMarkDefaults}
\title{helper function to provide line properties for marks}
\usage{
linesMarkDefaults()
}
\value{
a list object of parameters
}
\description{
helper function to provide line properties for marks
}
\note{
uses lineAttributes function
}
