% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{newDataDF}
\alias{newDataDF}
\title{creates a data.frame with the columns name, predictor,
 calculated response, lower & upper levels for the prediction}
\usage{
newDataDF(
  fit,
  weight = NULL,
  sampleTable,
  nameColumn = 1,
  predictionColumn = 2,
  predictionLevel = 0.95
)
}
\arguments{
\item{fit}{lm class object containg the fit info}

\item{weight}{default = NULL, otherwise must be integer to give weights to
the predictor values (see also function weightsVector)}

\item{sampleTable}{data.frame containing the data to be used for prediction,
must contain column with (sample-)names and a prediction/or column}

\item{nameColumn}{can be integer (column number) or character (column name)}

\item{predictionColumn}{can be integer (column number) or character
(column name) indicates the predictor column}

\item{predictionLevel}{default is 0.95, used for the prediction}
}
\value{
a data.frame
}
\description{
creates a data.frame with the columns name, predictor,
 calculated response, lower & upper levels for the prediction
}
\note{
use same weight as used for fit!
}
