% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{normalQQPlot}
\alias{normalQQPlot}
\title{creates a ggplot object showing a normal quantile plot with
 confidence intervals
 
 qqplot: https://github.com/aloy/qqplotr}
\usage{
normalQQPlot(
  data,
  column,
  removeNA = TRUE,
  sampleSize = NA,
  pointColor = "black",
  pointFill = "green",
  pointShape = 21,
  pointSize = 3,
  pointAlpha = 0.75,
  lineColor = "red",
  lineType = "solid",
  lineWidth = 1,
  lineAlpha = 1,
  title = NULL,
  distribution = "norm",
  distributionParameters = list(),
  bandType = c("pointwise", "boot", "ks", "ts")[3],
  bandFill = "blue",
  bandAlpha = 0.1,
  xAxis = TRUE,
  yAxis = TRUE,
  vertical = FALSE,
  xDefault = TRUE,
  xLimits = c(0, NA),
  yDefault = TRUE,
  yLimits = c(0, NA),
  xLabel = "Theoretical",
  yLabel = "Sample",
  ...
)
}
\arguments{
\item{data}{the data to be plotted, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which columns are to be used for the plot.
Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this will remove warning messages and errors}

\item{sampleSize}{allows to the use of a sample of the data to be used for
the boxplot. By default sampleSize = NA, in which case all data is used}

\item{pointColor}{defines the color of the border of the data points}

\item{pointFill}{defines the color of the data points themselves}

\item{pointShape}{shape of the data points}

\item{pointSize}{size of the data points}

\item{pointAlpha}{alpha ('see through' value) of the data points}

\item{lineColor}{color of the normal distribution 'line'}

\item{lineType}{type of the normal distribution 'line'}

\item{lineWidth}{width of the normal distribution 'line'}

\item{lineAlpha}{alpha ('see through' value) of the normal distribution
'line'}

\item{title}{sets title of graph}

\item{distribution}{character vector, default is 'norm', see
qqplotr::geom_qq_band for more info}

\item{distributionParameters}{list of additional parameters passed on to the
previously chosen distribution function, see 
qqplotr::geom_qq_band for more info}

\item{bandType}{sets the type of confidence bands calculations.
Default = "ks" (Kolmogorov-Smirnov test), for descriptions of the other
options ("pointwise","boot","ts") see ?geom_qq_band}

\item{bandFill}{sets the color of the confidence interval band}

\item{bandAlpha}{sets the alpha ('see through' value) of the confidence
interval band}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xDefault}{default is set to TRUE, together with xLimits, this can be
used to define the exact range of the X-axis}

\item{xLimits}{default = c(0,NA), together with xDefault, this can be
used to define the exact range of the Y-axis}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yDefault, this can be
used to define the exact range of the Y-axis}

\item{xLabel}{specifies the label on the x-axis (theoretical)}

\item{yLabel}{specifies the label on the y-axis (sample)}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}
}
\value{
a ggplot object
}
\description{
creates a ggplot object showing a normal quantile plot with
 confidence intervals
 
 qqplot: https://github.com/aloy/qqplotr
}
