% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{normalQuantilePlot}
\alias{normalQuantilePlot}
\title{creates a ggplot object showing a normal quantile plot}
\usage{
normalQuantilePlot(
  data,
  column,
  removeNA = TRUE,
  sampleSize = NA,
  pointColor = "black",
  pointFill = "green",
  pointShape = 21,
  pointSize = 3,
  pointAlpha = 0.75,
  lineColor = "red",
  lineType = "dotted",
  lineWidth = 1,
  lineAlpha = 1,
  title = NULL,
  xAxis = TRUE,
  yAxis = TRUE,
  vertical = FALSE,
  xDefault = FALSE,
  xLimits = c(0, NA),
  yDefault = TRUE,
  yLimits = c(0, NA),
  ...
)
}
\arguments{
\item{data}{the data to be plotted, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which columns are to be used for the plot.
Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this will remove warning messages and errors}

\item{sampleSize}{allows to the use of a sample of the data to be used for
the boxplot. By default sampleSize = NA, in which case all data is used}

\item{pointColor}{defines the color of the border of the data points}

\item{pointShape}{shape of the data points}

\item{pointSize}{size of the data points}

\item{lineColor}{color of the normal distribution 'line'}

\item{lineType}{type of the normal distribution 'line'}

\item{lineAlpha}{alpha ('see through' value) of the normal distribution
'line'}

\item{title}{sets title of graph}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xDefault}{default is set to TRUE, together with xLimits, this can be
used to define the exact range of the X-axis}

\item{xLimits}{default = c(0,NA), together with xDefault, this can be
used to define the exact range of the Y-axis}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yDefault, this can be
used to define the exact range of the Y-axis}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}

\item{pointfill}{defines the color of the data points themselves}

\item{pointsAlpha}{alpha ('see through' value) of the data points}

\item{lineWidh}{width of the normal distribution 'line'}
}
\value{
a ggplot object
}
\description{
creates a ggplot object showing a normal quantile plot
}
