% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{plotPlusMatrix}
\alias{plotPlusMatrix}
\title{function to generate an aligned set of (maximum 4) plots as a 2x2 matrix
 (2 columns, 2 rows)}
\usage{
plotPlusMatrix(
  sPlot,
  xPlot = clearPlot(),
  yPlot = clearPlot(),
  crossPlot = clearPlot(),
  widths = c(125, 875),
  heights = c(875, 125)
)
}
\arguments{
\item{sPlot}{ggplot object to be placed in top, right corner}

\item{xPlot}{ggplot object to be placed in bottom, right corner (along
x-axis of sPlot object)}

\item{yPlot}{ggplot object to be placed in top, left corner (along y-axis of
sPlot object)}

\item{crossPlot}{ggplot object to be placed in bottom, left corner}

\item{widths}{horizontal: two number (integer) vector specifying the amount
of the plot to be used for the plots (horizontally)}

\item{heights}{vertical: two number (integer) vector specifying the amount of
the plot to be used for the plots (vertically)}
}
\value{
a ggplot object
}
\description{
function to generate an aligned set of (maximum 4) plots as a 2x2 matrix
 (2 columns, 2 rows)
}
\note{
what is returned is grobtable (class gtable): this can be drawn via
 the grid.draw() function. The plot drawn before this will have to be cleared
 "manually" via clearPlot()
}
