% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{scatterPlot}
\alias{scatterPlot}
\title{generates a ggplot object of a scatterplot}
\usage{
scatterPlot(
  data,
  xColumn = 1,
  yColumn = 2,
  xTransform = NA,
  yTransform = NA,
  removeNA = FALSE,
  xLabel = ifelse(is.character(xColumn), xColumn, colnames(data)[xColumn]),
  yLabel = ifelse(is.character(yColumn), yColumn, colnames(data)[yColumn]),
  pointAlpha = 0.75,
  pointColor = "black",
  pointFill = "red",
  pointShape = 21,
  pointSize = 2,
  smoothLine = FALSE,
  smoothLineType = "solid",
  smoothLineColor = "black",
  smoothAlpha = 0.1,
  smoothFill = "lightblue",
  smoothWidth = 0.5,
  smoothOrientation = "x",
  smoothConfidence = FALSE,
  smoothMethod = NULL,
  title = paste(c(yLabel, " vs ", xLabel), collapse = ""),
  vertical = FALSE,
  xAxis = TRUE,
  xDefault = TRUE,
  xLimits = c(0, NA),
  xLog = FALSE,
  xSymmetric = FALSE,
  xSymmetricExpand = 0.05,
  xCentered = TRUE,
  xMedian = FALSE,
  xDeviations = 4,
  yAxis = TRUE,
  yDefault = TRUE,
  yLimits = c(0, NA),
  yLog = FALSE,
  ySymmetric = FALSE,
  ySymmetricExpand = 0.05,
  yCentered = TRUE,
  yMedian = FALSE,
  yDeviations = 4,
  gridLines = TRUE,
  gridLinesX = TRUE,
  gridLinesY = TRUE,
  ...
)
}
\arguments{
\item{data}{the data tp be plotted, data.frame or similar}

\item{xColumn}{specifies which column in the data argument contains the
x-data (refer to column via number or character vector (column name))}

\item{yColumn}{specifies which column in the data argument contains the
y-data (refer to column via number or character vector (column name))}

\item{xTransform}{specifies the function with which to transform the
x-data (if needed), use transformData() for this}

\item{yTransform}{specifies the function with which to transform the
y-data (if needed), use transformData() for this}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this has consquence that ROWS will be removed when using
multiple columns with data.frame's}

\item{xLabel}{defines x-axis label}

\item{yLabel}{defines y-axis label}

\item{pointAlpha}{alpha ('see through' value) of the data points}

\item{pointColor}{defines the color of the border of the data points}

\item{pointShape}{shape of the data points}

\item{pointSize}{size of the data points}

\item{smoothLine}{if TRUE then a smoothing line is drawn (via geom_smooth())}

\item{smoothLineType}{type of the smoothing line}

\item{smoothLineColor}{color of the smoothing line}

\item{smoothAlpha}{alpha ('see through' value) of the smoothing line}

\item{smoothFill}{specifies fill color pf tje smoothing confidence 'band'}

\item{smoothOrientation}{eihter "x" or "y", specifies what the smoothing
orientation should be. See ?ggplot2::geom_smooth for more information}

\item{smoothConfidence}{if TRUE then confidence 'band' is drawn around the
smoothing line}

\item{smoothMethod}{defines smoothin method. See ?ggplot2::geom_smooth for
more info}

\item{title}{specifies the title}

\item{vertical}{if TRUE then the plot is rotated 90 degrees (swap of X & Y)}

\item{xAxis}{if TRUE, the x-axis is properly draw with label, ticks etc}

\item{xDefault}{if default is set to FALSE, then together with xLimits, this
can be used to define the exact range of the x-axis}

\item{xLimits}{default = c(0,NA), together with xDefault, this can be
used to define the exact range of the x-axis}

\item{xLog}{if TRUE then logarithmic scale is used for the x-axis}

\item{xSymmetric}{if TRUE then the range of x-axis will be adjusted to be
equal on both the left and the right side of the center}

\item{xSymmetricExpand}{allows for padding around data (x-axis), 0.05 means
5% extra wide x-axis range}

\item{xCentered}{if TRUE, the plot will be 'cemtered' around the either the
mean or median x-value}

\item{xMedian}{if TRUE then median and mean absolute deviation (mad) are
used for centering the plot along the x-axis; if FALSE then the mean and the
standard deviation are used}

\item{xDeviations}{defines how many deviations the range of the x-axis may
differ from the mean or median. Range will be either (median-xDeviations*mad
,median+xDeviations**mad) or (mean - xDeviations*sd,mean + xDeviations*sd)}

\item{yAxis}{if TRUE, the y-axis is properly draw with label, ticks etc}

\item{yDefault}{if default is set to FALSE, then together with yLimits, this
can be used to define the exact range of the y-axis}

\item{yLimits}{default = c(0,NA), together with yDefault, this can be
used to define the exact range of the y-axis}

\item{yLog}{if TRUE then logarithmic scale is used for the y-axis}

\item{ySymmetric}{if TRUE then the range of y-axis will be adjusted to be
equal on both the left and the right side of the center}

\item{ySymmetricExpand}{allows for padding around data (y-axis), 0.05 means
5% extra wide y-axis range}

\item{yCentered}{if TRUE, the plot will be 'cemtered' around the either the
mean or median y-value}

\item{yMedian}{if TRUE then median and mean absolute deviation (mad) are
used for centering the plot along the y-axis; if FALSE then the mean and the
standard deviation are used}

\item{yDeviations}{defines how many deviations the range of the x-axis may
differ from the mean or median. Range will be either (median-xDeviations*mad
,median+xDeviations**mad) or (mean - xDeviations*sd,mean + xDeviations*sd)}

\item{gridLines}{if TRUE then gridlines are shown}

\item{gridLinesX}{if TRUE then vertical gridlines are shown (set gridLines to
FALSE when using this)}

\item{gridLinesY}{if TRUE then horizontal gridlines are shown (set gridLines
to FALSE when using this)}

\item{pointfill}{defines the color of the data points themselves}

\item{smoothWidh}{width of the smoothing nline}
}
\value{
a ggplot object
}
\description{
generates a ggplot object of a scatterplot
}
