% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{scatterPlotPlusH3}
\alias{scatterPlotPlusH3}
\title{function to generate an aligned set of (maximum 3) plots as a 1x3 matrix
 (3 columns, 1 row)}
\usage{
scatterPlotPlusH3(
  sPlot,
  yLeft = clearPlot(),
  yRight = clearPlot(),
  widths = c(150, 500, 350)
)
}
\arguments{
\item{sPlot}{ggplot object to be placed in the middle}

\item{yLeft}{ggplot object to be placed on the left (along y-axis of sPlot)}

\item{widths}{horizontal: two number (integer) vector specifying the amount
of the plot to be used for the plots (horizontally)}

\item{yright}{ggplot object to be placed on the right (along y-axis of sPlot)}
}
\value{
a ggplot object
}
\description{
function to generate an aligned set of (maximum 3) plots as a 1x3 matrix
 (3 columns, 1 row)
}
\note{
what is returned is grobtable (class gtable): this can be drawn via
 the grid.draw() function. The plot drawn before this will have to be cleared
 "manually" via clearPlot()
}
