% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{statBoxPlotMultiTable}
\alias{statBoxPlotMultiTable}
\title{takes a lists of data.frame's and generates a boxplot using data from the
 different data.frame's in the list as elements in the boxplot}
\usage{
statBoxPlotMultiTable(
  data,
  idColumn = 1,
  varColumn = 2,
  sortIDs = FALSE,
  sortDescending = FALSE,
  sampleSize = NA,
  removeNA = FALSE,
  variableName = "variable",
  valueName = "value",
  outlineColor = "black",
  fillColor = "red",
  jitter = 0.05,
  alpha = 0.5,
  size = 3,
  shape = 16,
  jitterFill = "black",
  whiskerWidth = 0.5,
  boxWidth = 0.5,
  vertical = FALSE,
  xAxis = TRUE,
  yAxis = TRUE,
  xDiscrete = TRUE,
  yDefault = TRUE,
  yLimits = c(0, NA),
  xLabel = "",
  yLabel = "",
  title = "",
  showMean = TRUE,
  meanShape = 23,
  meanColor = "black",
  meanFill = "orange",
  meanSize = 5,
  returnData = FALSE,
  ...
)
}
\arguments{
\item{data}{the data to be plotted. Must be a list of data.frame's with each
at least the same idColumn & varColumn present}

\item{idColumn}{defines which column is to be used for x-axis of the boxplot.
Can be integer or character (column name)}

\item{varColumn}{defines which column is to be used for the y-axis of the
boxplot}

\item{sortIDs}{if TRUE then data in the idColumn will be sorted}

\item{sortDescending}{if TRUE then the data will be sorted in a descending}

\item{sampleSize}{allows to the use of a sample of the data to be used for
the boxplot. By default sampleSize = NA, in which case all data is used}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this will remove warning messages and errors}

\item{variableName}{sets the 'combined' name of the columns, 
must be a single word}

\item{outlineColor}{defines the color of the line around the box}

\item{fillColor}{defines the color of the boxes themselves. @Note: if the 
number of colors does not match the number of columns then ggplot2 default
colors will be used}

\item{jitter}{if NA, then the data points will not be shown (only outliers!),
otherwise it adds a random value to the x-values of the data points plotted.
Note: If set to 0 then they will be located on a straight line}

\item{alpha}{alpha ('see through' value) of the data (jitter) points}

\item{size}{size of the data (jitter) points}

\item{shape}{shape of the data (default = 16), see vignette
ggplot2::ggplot2-specs}

\item{jitterFill}{defines color of the jitter (single color!)}

\item{whiskerWidth}{defines the width of the whiskers (0-1)}

\item{boxWidth}{defines the width of the box (0-1)}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLabel}{set y-axis title}

\item{title}{sets title of graph}

\item{showMean}{defines if the mean value of the data should be shown}

\item{meanShape}{shape of the mean symbol (default = 23)}

\item{meanColor}{color of the line around the mean symbol}

\item{meanFill}{fill color of the shape of the mean symbol}

\item{meanSize}{size of the mean symbol}

\item{returnData}{if TRUE then a list with 2 elements is returned. The first
element is the data.frame used to generate the graph and the second element
is the graph itself}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}

\item{showLegend}{defines if the legend is to be shown or not}

\item{legend.position}{defines where a legend is to be placed}
}
\value{
a ggplot object or a list
}
\description{
takes a lists of data.frame's and generates a boxplot using data from the
 different data.frame's in the list as elements in the boxplot
}
\note{
the data in the idColumns & varColumns are essentially x,y paurs, this
 function combines all the x's from the different data.frame's in the list
 and then makes a boxplot. Example code:
 l <- list()
 l[[1]] <- data.frame(name = c("P1","P2","P3","P4"), value = c(1,2,3,2))
 l[[2]] <- data.frame(name = c("P1","P2","P3","P4"), value = c(4,5,10,5.5))
 l[[3]] <- data.frame(name = c("P1","P2","P3","P4"), value = c(10,2,23,5))
 l[[4]] <- data.frame(name = c("P1","P2","P4"), value = c(10,2,6.7))
 statBoxPlotMultiTable(l, idColumn = "name",
                        varColumn = "value", returnData = T)

box itself:
 bottom = 25% quantile, top = 75% quantile, middle = 50% quantile (median),
 lower whisker = 25% quantile + 1.5*IQR,
 upper whisker= 75% quantile + 1.5*IQR 
 IQR = (75% quantile) - (25% quantile) (Inter Quantile Range)
}
