% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{statBoxPlotSingle}
\alias{statBoxPlotSingle}
\title{creates a ggplot object showing a boxplot}
\usage{
statBoxPlotSingle(
  data,
  column = 1,
  removeNA = TRUE,
  sampleSize = NA,
  outlineColor = "black",
  fillColor = "red",
  jitter = 0.05,
  alpha = 0.5,
  size = 3,
  shape = 16,
  whiskerWidth = 0.5,
  boxWidth = 0.5,
  xAxis = FALSE,
  yAxis = TRUE,
  yLabel = ifelse(!is.Class(data, "data.frame"), NA, ifelse(is.character(column),
    paste(column, collapse = ""), paste(colnames(data)[column], collapse = ", "))),
  title = NA,
  vertical = FALSE,
  xDefault = FALSE,
  xLimits = c(-0.75, 0.75),
  yDefault = TRUE,
  yLimits = c(0, NA),
  showMean = TRUE,
  meanShape = 23,
  meanColor = "black",
  meanFill = "orange",
  meanSize = 5,
  ...
)
}
\arguments{
\item{data}{the data to be plotted, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which column is to be used for the boxplot.
Can be integer or character (column name), note that if both (character)
column and yLabel are defined, column is used as label for the Y-axis}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this will remove warning messages and errors}

\item{sampleSize}{allows to the use of a sample of the data to be used for
the boxplot. By default sampleSize = NA, in which case all data is used}

\item{outlineColor}{defines the color of the line around the box}

\item{fillColor}{defines the color of the box itself}

\item{jitter}{if NA, then the data points will not be shown (only outliers!),
otherwise it adds a random value to the x-values of the data points plotted.
Note: If set to 0 then they will be located on a straight line}

\item{alpha}{alpha ('see through' value) of the data points}

\item{size}{size of the data points}

\item{shape}{shape of the datapoints (default = 16), see vignette
ggplot2::ggplot2-specs}

\item{whiskerWidth}{defines the width of the whiskers (0-1)}

\item{boxWidth}{defines the width of the box (0-1)}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{yLabel}{set y-axis title}

\item{title}{sets title of graph}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xDefault}{default is set to FALSE, together with xLimits this can be
used to tweak the positioning and with of the box}

\item{xLimits}{default = c(-0.75,0.75), together with xDefault this can be
used to tweak the positioning and with of the box}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{showMean}{defines if the mean value of the data should be shown}

\item{meanShape}{shape of the mean symbol (default = 23)}

\item{meanColor}{color of the line around the mean symbol}

\item{meanFill}{fill color of the shape of the mean symbol}

\item{meanSize}{size of the mean symbol}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}
}
\value{
a ggplot object
}
\description{
creates a ggplot object showing a boxplot
}
\note{
box itself:
 bottom = 25% quantile, top = 75% quantile, middle = 50% quantile (median),
 lower whisker = 25% quantile + 1.5*IQR,
 upper whisker= 75% quantile + 1.5*IQR 
 IQR = (75% quantile) - (25% quantile) (Inter Quantile Range)
}
