% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{statHist}
\alias{statHist}
\title{creates a ggplot object showing a histogram}
\usage{
statHist(
  data,
  column = 1,
  binwidth = NULL,
  bins = NULL,
  statCount = FALSE,
  variableName = "variable",
  removeNA = TRUE,
  outlineColor = "white",
  fillColor = "red",
  xLabel = ifelse(!is.Class(data, "data.frame"), NA, ifelse(is.character(column),
    paste(column, collapse = ", "), paste(colnames(data)[column], collapse = ", "))),
  yLabel = "Frequency",
  title = NA,
  titleDefault = paste(c("Histogram of ", ifelse(!is.character(column), "", column)),
    collapse = ""),
  xAxis = TRUE,
  yAxis = TRUE,
  xDefault = TRUE,
  xLimits = c(0, NA),
  xSymmetric = FALSE,
  xSymmetricExpand = 0.05,
  xCentered = FALSE,
  xMedian = FALSE,
  xDeviations = 4,
  showLegend = TRUE,
  legend.position = "bottom",
  vertical = FALSE,
  ...
)
}
\arguments{
\item{data}{the data tp be plotted, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
defines which column(s) is/are to be used}

\item{column}{defines which column(s) is/are to be used for the histogram.
Can be integer or character (column name(s))}

\item{binwidth}{defines width of the 'bins' of the histogram, if NULL
(default), then it will be set automatically (with a warning). This setting
is ignored in case of statCount is set tp TRUE}

\item{bins}{defines the number of 'bins' of the histogram, overriden by
binwidth}

\item{statCount}{set to TRUE if the data is not numerical}

\item{variableName}{sets the 'combined' name of the columns (IF there is more
than one!)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this has consquence that ROWS will be removed when using
multiple columns with data.frame's}

\item{outlineColor}{defines the color of the line around the bars}

\item{fillColor}{defines the color of the bars themselves. If a multi-column
data.frame is plotted, the same number as the number of columns used should
be used. If not the same number, then the graph will revert to default
colors of ggplot}

\item{xLabel}{sets x-axis title}

\item{yLabel}{set y-axos title}

\item{title}{sets title of graph, if NA then the titleDefault will be used}

\item{titleDefault}{will be combined with the xLabel to be used as title if
title == NA}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{xDefault}{this defines if default x-sxis limits should be used or not,
see also graphAdjust() for info}

\item{xLimits}{default = c(0,NA), together with xDefault, this can be
used to define the exact range of the x-axis}

\item{xSymmetric}{if TRUE then the range of x-axis will be adjusted to be
equal on both the left and the right side of the center}

\item{xSymmetricExpand}{allows for padding around data (x-axis), 0.05 means
5 percent extra wide x-axis range}

\item{xCentered}{if TRUE, the plot will be 'cemtered' around the either the
mean or median x-value}

\item{xMedian}{if TRUE then median and mean absolute deviation (mad) are
used for centering the plot along the x-axis; if FALSE then the mean and the
standard deviation are used}

\item{xDeviations}{defines how many deviations the range of the x-axis may
differ from the mean or median. Range will be either (median-xDeviations*mad
,median+xDeviations**mad) or (mean - xDeviations*sd,mean + xDeviations*sd)}

\item{showLegend}{defines if the legend is to be shown or not}

\item{legend.position}{defines where a legend is to be placed}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}
}
\value{
a ggplot object
}
\description{
creates a ggplot object showing a histogram
}
