% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{stdRecoveryLimits}
\alias{stdRecoveryLimits}
\title{helper function to provide default limits for recoveryPlots}
\usage{
stdRecoveryLimits()
}
\value{
data.frame of the attributes
}
\description{
helper function to provide default limits for recoveryPlots
}
