% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{stdResidualLimits}
\alias{stdResidualLimits}
\title{wrapper function for (default) limit line attributes}
\usage{
stdResidualLimits()
}
\value{
data.frame of the attributes
}
\description{
wrapper function for (default) limit line attributes
}
