% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools.R
\name{tableQuantiles}
\alias{tableQuantiles}
\title{returns a data.frame with the specified quantile limits of data}
\usage{
tableQuantiles(
  data,
  column,
  removeNA = TRUE,
  limits = c(1, 0.995, 0.975, 0.9, 0.75, 0.5, 0.25, 0.1, 0.025, 0.5, 0),
  label = NA,
  quantilePercentage = TRUE
)
}
\arguments{
\item{data}{the data to be used, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which column is to be used for calculating the
quantiles. Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to calculating the quantiles. @note this will remove warning messages and
errors}

\item{limits}{a numeric vector specifying the quantile limits to be
calculated. @note these values should be between 0 and 1}

\item{label}{defines the name of the calculated limits column. Under 'normal'
circumstances this would be the column name of the data}

\item{quantilePercentage}{boolean defining whether the percentile column 
output should percentages of the values specified in limits parameter}
}
\value{
a data.frame with two columns: quanitle and the limits calculated
 from the data
}
\description{
returns a data.frame with the specified quantile limits of data
}
