% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools.R
\name{tableSummary}
\alias{tableSummary}
\title{returns a statistics summary table of the specified data (mean and
 associated statistical measures)}
\usage{
tableSummary(data, column, removeNA = TRUE, alpha = 0.05)
}
\arguments{
\item{data}{the data to be used, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which column is to be used for calculating the
statistics. Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to the calculations. @note this will remove warning messages and errors}

\item{alpha}{specifies the confidence level to be used. @note this value
should be between 0 and 1}
}
\value{
a data.frame with two columns: statistic name and the calculated
 statistic itself
}
\description{
returns a statistics summary table of the specified data (mean and
 associated statistical measures)
}
