% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools.R
\name{toleranceInterval}
\alias{toleranceInterval}
\title{returns a data.frame with 2-sided tolerance intervals for the data
 distributed according to normal distribution}
\usage{
toleranceInterval(
  data,
  column,
  removeNA = TRUE,
  alpha = 0.05,
  proportion = 0.95
)
}
\arguments{
\item{data}{the data to be used, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which column is to be used for calculating the
quantiles. Can be integer or character (column name)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to calculating the quantiles. @note this will remove warning messages and
errors}

\item{alpha}{(1-confidence level), specifies the confidence level required}

\item{proportion}{specifies proportion of the population}
}
\value{
a data.frame with two columns: statistic and the values calculated
 from the data
}
\description{
returns a data.frame with 2-sided tolerance intervals for the data
 distributed according to normal distribution
}
\note{
tolerance interval = the range that covers a specified proportion of
 the population with a given level of confidence (1-alpha). It expresses the
 range where individual values from the population are expect to be
}
