% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{toNumeric}
\alias{toNumeric}
\title{support function: changes character vector of numbers into numerical vector
 (and adds NA's when appropriate)}
\usage{
toNumeric(chrs = "N/A")
}
\arguments{
\item{chrs}{character vector of numbers in 'text' format}
}
\value{
numeric vector
}
\description{
support function: changes character vector of numbers into numerical vector
 (and adds NA's when appropriate)
}
