% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{transform01Logical}
\alias{transform01Logical}
\title{takes logical data and replaces it with an alternative, eg "Yes" and "No"}
\usage{
transform01Logical(data, alternative = c("No", "Yes"), na.alternative = " ")
}
\arguments{
\item{data}{data to be transformed, should be logicalm but can be 1's and 0's
(internally these will be transformed into TRUE's and FALSE's respectively)}

\item{alternative}{specifies what to transform TRUE (element 2) or FALSE
(element 1) into}

\item{na.alternative}{specifies what to use if NA is encountered

 transform01Logical(c(T,F,F,T))
 transform01Logical(c(1,0,0,1))}
}
\value{
data with alternatives in stead of TRUE/FALSE
}
\description{
takes logical data and replaces it with an alternative, eg "Yes" and "No"
}
