% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{transformData}
\alias{transformData}
\title{helper function: function factory to generalize data transformations which
 may or may not be needed for volcanoPlot()}
\usage{
transformData(transformationFormula = "data")
}
\arguments{
\item{transformationFormula}{character vector that specifies the formula
to be used on the argument "data". As an example "log2(data)"}
}
\value{
a function that takes the argument "data" and transforms it
}
\description{
helper function: function factory to generalize data transformations which
 may or may not be needed for volcanoPlot()
}
