% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{transformNA}
\alias{transformNA}
\title{Replaces all NA's in data (data.frame, list, etc) with an alternative}
\usage{
transformNA(data, na.alternative = "-")
}
\arguments{
\item{data}{data.frame, list or ... that contains NA's which need to be
replaced}

\item{na.alternative}{the object with which NA's need to be replace. Be aware
that a list or column of a data.frame may change class when taking this
action!}
}
\value{
data with replaced NA's
}
\description{
Replaces all NA's in data (data.frame, list, etc) with an alternative
}
