% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{transformToDateString}
\alias{transformToDateString}
\title{takes a 'Date' object or an integer and converts it to a character vector of
 a specified format. Also changes NA's into a specified alternative}
\usage{
transformToDateString(
  aDate,
  formatString = "\%d/\%m/\%Y",
  origin = "1970-01-01",
  na.alternative = NA
)
}
\arguments{
\item{aDate}{date vector or integer vector (days since origin)}

\item{formatString}{defines the formatting of the date vector, internally the
base::as.Date function is used (via integerToDate)}

\item{origin}{character vector specifying the origin date (only used if aDate
argument is an integer)}

\item{na.alternative}{specifies what to use for NA elements in the aDate
 argument

transformData(transformationFormula = "transformToDate(data, na.alternative = '-')")
transformToDate(c(Sys.Date(),NA, Sys.Date()+1))
transformToDate(c(Sys.Date(),NA, Sys.Date()+1), na.alternative = "-")}
}
\value{
character vector
}
\description{
takes a 'Date' object or an integer and converts it to a character vector of
 a specified format. Also changes NA's into a specified alternative
}
