% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{vectorToBlob}
\alias{vectorToBlob}
\title{Converts a numeric, integer or character vector (multiple elements) to a
 sort of raw vector. (actually a string representation of a raw vector).
 Note: also compresses data if required.}
\usage{
vectorToBlob(
  dataVector = NA,
  collapseChar = ";",
  formatNumbers = FALSE,
  type = "gzip",
  ...
)
}
\arguments{
\item{dataVector}{the data to be converted: integer, numeric or character.
Other data types have not been tested}

\item{collapseChar}{the character to be used as seperator in the end result.
Note: this seperator should not be present in the dataVector (especially
after transformation) as it will give problems when using the DBToVector
function.}

\item{formatNumbers}{boolean specifying whether the dataVector should be
formatted using the base::format function}

\item{type}{character string, the type of compression. See ?memCompress for
details}

\item{...}{further arguments to be passed onto to base::format, see
?base::format for more information. Ignored when formatNumbers == FALSE}
}
\value{
a character vector
}
\description{
Converts a numeric, integer or character vector (multiple elements) to a
 sort of raw vector. (actually a string representation of a raw vector).
 Note: also compresses data if required.
}
