% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{vectorToCode}
\alias{vectorToCode}
\title{programming tool to convert a vector in memory to code for inclusion
 into programs.}
\usage{
vectorToCode(vectorData, includeParenthesis = TRUE, trim = NA, ...)
}
\arguments{
\item{vectorData}{can be vector of class character, integer, logical,
numeric, etc. Note that custom data types may not work properly.}

\item{includeParenthesis}{boolean parameter defining if 'c()' should be
included around the character vector returned.}

\item{trim}{if NA (default) nothing is done, else should be value "both",
"left" or "right". It will then cause all vectors to be white space trimmed
on the specified site. See ?stringr::str_trim for more details. Note that
trimming is done after formatting, so may undo settings like wide = x'}

\item{...}{is used to transfer settings to the base format() function. See
?format for more info}
}
\value{
a character vector.
}
\description{
programming tool to convert a vector in memory to code for inclusion
 into programs.
}
