% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{vectorToDB}
\alias{vectorToDB}
\title{Converts a numeric, integer or character vector (multiple elements) to a
 single element character vector (a string)}
\usage{
vectorToDB(
  dataVector = NA,
  collapseChar = ";",
  formatNumbers = FALSE,
  removeNA = FALSE,
  ...
)
}
\arguments{
\item{dataVector}{the data to be converted: integer, numeric, character or
logical. Other data types have not been tested}

\item{collapseChar}{the character to be used as seperator in the end result.
Note: this seperator should not be present in the dataVector (especially
after transformation) as it will give problems when using the DBToVector
function.}

\item{formatNumbers}{boolean specifying whether the dataVector should be
formatted using the base::format function}

\item{removeNA}{if TRUE then all NA values are removed before the result is
transformed into a single element character vector}

\item{...}{further arguments to be passed onto to base::format, see
?base::format for more information. Ignored when formatNumbers == FALSE}
}
\value{
a single element character vector
}
\description{
Converts a numeric, integer or character vector (multiple elements) to a
 single element character vector (a string)
}
