% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{volcanoLineMarkerDedaults}
\alias{volcanoLineMarkerDedaults}
\title{helper function that specifies the marker lines to add to the volcanoplot()}
\usage{
volcanoLineMarkerDedaults()
}
\value{
a list of values for the x-axis (vlines) and the y-axis (hlines)
}
\description{
helper function that specifies the marker lines to add to the volcanoplot()
}
