% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{volcanoMarkerAttributesDefaults}
\alias{volcanoMarkerAttributesDefaults}
\title{helper function that specifies the marker line attributes to add to the
 volcanoplot()}
\usage{
volcanoMarkerAttributesDefaults()
}
\value{
a list of values for the x-axis markerlines (vlinesAttributes) and
 the y-axis markerlines (hlinesAttributes)
}
\description{
helper function that specifies the marker line attributes to add to the
 volcanoplot()
}
