% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{generates a volcano plot of a data.frame with quantification data & 
 statisitcal significance data}
\usage{
volcanoPlot(
  data,
  quantColumn = 1,
  statColumn = 2,
  quantTransform = transformData("log2(data)"),
  statTransform = transformData("-log10(data)"),
  xLabel = ifelse(is.character(quantColumn), paste(c("Log2 ", quantColumn), collapse =
    ""), paste(c("Log2 ", colnames(data)[quantColumn]), collapse = "")),
  yLabel = ifelse(is.character(statColumn), paste(c("-Log10 ", statColumn), collapse =
    ""), paste(c("-Log10 ", colnames(data)[statColumn]), collapse = "")),
  xCutoffs = c(0.5, 2),
  yCutoff = 0.05,
  pointColor = "black",
  significantPointColor = "red",
  pointShape = 21,
  pointSize = 3,
  pointAlpha = 0.75,
  xDefault = TRUE,
  yDefault = TRUE,
  xLimits = c(0, NA),
  yLimits = c(0, NA),
  xExpand = ggplot2::expansion(mult = 0, add = 0),
  yExpand = ggplot2::expansion(mult = 0, add = 0),
  xSymmetric = FALSE,
  xSymmetricExpand = 0.05,
  xCentered = FALSE,
  xMedian = FALSE,
  xDeviations = 4,
  xLabelFormat = ggplot2::waiver(),
  yLabelFormat = ggplot2::waiver(),
  vertical = FALSE,
  title = "",
  gridLines = TRUE,
  volcanoLineMarkers = volcanoLineMarkerDedaults(),
  volcanoLineMarkerAttributes = volcanoMarkerAttributesDefaults(),
  returnData = FALSE,
  significanceColumnName = "",
  removeNonSignificantData = TRUE,
  identifierColumn = NA,
  ...
)
}
\arguments{
\item{data}{data.frame with at least two columns}

\item{quantColumn}{specifies which column in the data argument contains the
quantification data (can be number or character vector)}

\item{statColumn}{specifies which column in the data argument contains the
statistical significance data (can be number or character vector). Usually
these values are p-values from eg a t-test}

\item{quantTransform}{specifies the function with which to transform the
quantColumn data (if needed)}

\item{statTransform}{specifies the function with which to transform the
stattColumn data (if needed)}

\item{xLabel}{defines x-axis label}

\item{yLabel}{defines y-axis label}

\item{xCutoffs}{minimum and maximum value of the quantColumn data
(before transformation). Anything outside these values and lower than
yCutoff will be marked as significant}

\item{yCutoff}{minimum value of the statColumn data. any value lower is
marked as significant (usually p-value < 0.05)}

\item{pointColor}{defines the color of the border of the data points}

\item{significantPointColor}{color of the data points that lie outside the
xCutoff values and below the yCutoff value}

\item{pointShape}{shape of the data points}

\item{pointSize}{size of the data points}

\item{pointAlpha}{alpha ('see through' value) of the data points}

\item{xDefault}{default is set to TRUE, together with xLimits, this can be
used to define the exact range of the X-axis}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{xLimits}{default = c(0,NA), together with xDefault, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yDefault, this can be
used to define the exact range of the Y-axis}

\item{xExpand}{allows for padding around data (x-axis),
see ?ggplot2::expansion for proper explanation}

\item{yExpand}{allows for padding around data (y-axis),
see ?ggplot2::expansion for proper explanation}

\item{xSymmetric}{if TRUE then the range of x-axis will be adjusted to be
equal on both the left and the right side of the center (transformed x = 0)}

\item{xSymmetricExpand}{allows for padding around data (x-axis), 0.05 means
5 percent (pre-transformation) extra wide x-axis range}

\item{xCentered}{if TRUE, the plot will be 'cemtered' around the either the
mean or median x-value}

\item{xMedian}{if TRUE then median and mean absolute deviation (mad) are
used for centering the plot along the x-axis; if FALSE then the mean and tge
standard deviation are used}

\item{xDeviations}{defines how many deviations the range of the x-axis may
differ from the mean or median. Range will be either (median-xDeviations*mad
,median+xDeviations**mad) or (mean - xDeviations*sd,mean + xDeviations*sd)}

\item{xLabelFormat}{allows defining how the x-axis ticks etc are displayed
(see formatDigits() & formatScientificDigits())}

\item{yLabelFormat}{allows defining how the y-axis ticks etc are displayed
(see formatDigits() & formatScientificDigits())}

\item{vertical}{if TRUE then plot will be rotated 90 degrees}

\item{title}{specifies the title}

\item{gridLines}{if TRUE then gridlines are shown}

\item{volcanoLineMarkers}{defines where to place marker lines in the plot}

\item{volcanoLineMarkerAttributes}{defines the attributes of the marker lines}

\item{returnData}{if TRUE then a list with 2 elements is returned. The first
element is the data.frame used to generate the graph and the second element
is the graph itself}

\item{significanceColumnName}{name to give to the column specifying if a
row is significant (TRUE) or (FALSE). If not specified, then the name
"significant" will be used}

\item{removeNonSignificantData}{if TRUE, no non-significant data will be
returned (usually a smaller data.frame to return). If FALSE then all data
will be returned}

\item{identifierColumn}{specifies which columns from data (names -> character
vector) should be included in the data.frame returned}

\item{...}{can be used to pass on other arguments to graphAdjust()}
}
\value{
a ggplot object or a list
}
\description{
generates a volcano plot of a data.frame with quantification data & 
 statisitcal significance data
}
\note{
quantColumn is the x-axis, the statColumn is the y-axis
}
