% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{weightsFromResidualsVector}
\alias{weightsFromResidualsVector}
\title{generates vector of weighted values where the weights are calculated from
 the residuals coming from a fit}
\usage{
weightsFromResidualsVector(fit)
}
\arguments{
\item{fit}{model object (eg coming from lm() or fitLM())}
}
\value{
a vector of weighted values (or NULL)
}
\description{
generates vector of weighted values where the weights are calculated from
 the residuals coming from a fit
}
