% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicR.R
\name{legendDefinition}
\alias{legendDefinition}
\title{helper function for the function customLegend():
 generates a data.frame specifying the legend to be generated}
\usage{
legendDefinition(labels = NA, colors = NA, fills = NA, shapes = NA, sizes = NA)
}
\description{
helper function for the function customLegend():
 generates a data.frame specifying the legend to be generated
}
