% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{scatterBlandAltman}
\alias{scatterBlandAltman}
\title{generates a variant of the scatterplot when one needs to compare two sets
 of data (x & y): Bland-Altman / Tukey mean-difference plot
 https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4470095/pdf/bm-25-141.pdf
 https://en.wikipedia.org/wiki/Bland%E2%80%93Altman_plot}
\usage{
scatterBlandAltman(
  data,
  xColumn = 1,
  yColumn = 2,
  logTransform = FALSE,
  log2Transform = TRUE,
  removeNA = FALSE,
  xLabel = ifelse(logTransform, ifelse(log2Transform, "Average of log2(x,y)",
    "Average of log10(x,y)"), "Average of c(x,y)"),
  yLabel = ifelse(logTransform, ifelse(log2Transform, "Difference of log2(x,y)",
    "Difference of log10(x,y)"), "Difference of c(x,y)"),
  title = paste(c("Bland-Altman plot/Tukey mean-difference plot: ", yLabel, " vs ",
    xLabel), collapse = ""),
  ...
)
}
\arguments{
\item{data}{the data tp be plotted, data.frame or similar}

\item{xColumn}{specifies which column in the data argument contains the
x-data (refer to column via number or character vector (column name))}

\item{yColumn}{specifies which column in the data argument contains the
y-data (refer to column via number or character vector (column name))}

\item{logTransform}{do logaithmic transformation o/t data}

\item{log2Transform}{if TRUE do log2 transformation o/t data in stead of
log10 (if logTransform = TRUE!)}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. Note: this has consquence that ROWS will be removed when using
multiple columns with data.frame's}

\item{xLabel}{defines x-axis label}

\item{yLabel}{defines y-axis label}

\item{title}{specifies the title}
}
\value{
a ggplot object
}
\description{
generates a variant of the scatterplot when one needs to compare two sets
 of data (x & y): Bland-Altman / Tukey mean-difference plot
 https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4470095/pdf/bm-25-141.pdf
 https://en.wikipedia.org/wiki/Bland%E2%80%93Altman_plot
}
\note{
work in progress...
}
