% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalStat.R
\name{statViolinPlotMultiple}
\alias{statViolinPlotMultiple}
\title{creates a ggplot object showing a violin plot of multiple columns}
\usage{
statViolinPlotMultiple(
  data,
  column = 1:ncol(data),
  sampleSize = NA,
  removeNA = TRUE,
  variableName = "variable",
  outlineColor = "black",
  fillColor = NA,
  outlineSize = 0.5,
  outlineType = "solid",
  violinAlpha = 1,
  scale = c("area", "count", "width")[1],
  trim = TRUE,
  bandwidth = 1,
  quantiles = NULL,
  jitter = 0.05,
  alpha = 0.5,
  size = 3,
  shape = 16,
  jitterFill = "black",
  vertical = FALSE,
  xAxis = TRUE,
  yAxis = TRUE,
  yDefault = TRUE,
  yLimits = c(0, NA),
  xLabel = "",
  yLabel = "",
  title = "",
  showLegend = TRUE,
  legend.position = "bottom",
  ...
)
}
\arguments{
\item{data}{the data to be plotted, can be a numeric/character/etc vector or
data.frame like (or tibble etc). If it is data.frame or similar the column
argument defines which column is to be used}

\item{column}{defines which columns are to be used for the boxplot.
Can be integer or character (column name), note that if both (character)
column and yLabel are defined, column is used as label for the Y-axis. If
not defined, then all columns of the data.frame will be used.}

\item{sampleSize}{allows to the use of a sample of the data to be used for
the boxplot. By default sampleSize = NA, in which case all data is used}

\item{removeNA}{if TRUE, the NA 'values' in the vector will be removed prior
to plotting. @note this will remove warning messages and errors}

\item{variableName}{sets the 'combined' name of the columns (IF there is more
than one!)}

\item{outlineColor}{defines the color of the line around the 'violin'}

\item{fillColor}{defines the color of the boxes themselves. @Note: if the 
number of colors does not match the number of columns then ggplot2 default
colors will be used}

\item{outlineSize}{defines the width of the line around the 'violin'}

\item{outlineType}{defines the linetype of the line around the 'violin'}

\item{violinAlpha}{defines the alpha ('see through' value) the 'violin'}

\item{scale}{default = "area", other valid values are "count" and "width".
if "area" (default), all violins have the same area (before trimming the
tails). If "count", areas are scaled proportionally to the number of
 observations. If "width", all violins have the same maximum width.
 @note taken from ?geom_violin, there is no sense in setting this argument
 for a single violin plot}

\item{trim}{If TRUE (default), trim the tails of the violins to the range of
the data. If FALSE, don't trim the tails. @note taken from ?geom_violin}

\item{bandwidth}{defines the 'adjust' parameter, which is the adjustment of
smoothing bandwidth. 1/2 means half of default bandwidth --> makes the
outline more 'jagged'}

\item{quantiles}{draws lines at the specified quantiles, eg c(0.25,0.5,0.25).
If NULL, nothing is drawn}

\item{jitter}{if NA, then the data points will not be shown (only outliers!),
otherwise it adds a random value to the x-values of the data points plotted.
Note: If set to 0 then they will be located on a straight line}

\item{alpha}{alpha ('see through' value) of the data points}

\item{size}{size of the data points}

\item{shape}{shape of the datapoints (default = 16), see vignette
ggplot2::ggplot2-specs}

\item{jitterFill}{fill color of the data points}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xAxis}{defines if the x-axis is shown}

\item{yAxis}{defines if the x-axis is shown}

\item{yDefault}{default is set to TRUE, together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{yLimits}{default = c(0,NA), together with yLimits, this can be
used to define the exact range of the Y-axis}

\item{xLabel}{set x-axis title}

\item{yLabel}{set y-axis title}

\item{title}{sets title of graph}

\item{showLegend}{defines if the legend is to be shown or not}

\item{legend.position}{defines where a legend is to be placed}

\item{...}{can be used to pass on other arguments to graphAdjust()
(like xLimits, xExpand, etc)}
}
\value{
a ggplot object
}
\description{
creates a ggplot object showing a violin plot of multiple columns
}
